# src/main/openapi/bus-ticket-booking-app.yml
openapi: 3.1.0
info:
  title: Bus Ticket Booking API
  description: API for managing bus ticket booking application
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearerAuth: [ ]
...
paths:
...
  /api/v1/scheduled-trips:
    post:
      tags:
        - Scheduled Trip Management
      summary: Create Scheduled Trip
      description: Create a new scheduled trip based on the provided request.
      operationId: CreateScheduledTrip
      requestBody:
        description: Scheduled trip request containing details for creating a new
          scheduled trip
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduledTripRequest"
        required: true
      responses:
        "201":
          description: Successfully created scheduled trips
          content:
            application/json:
              schema:
                type: string
                example: Successfully created 5 scheduled trips
        "400":
          description: Invalid request data
          content:
            application/json:
              schema:
                type: string
                example: Invalid scheduled trip request
        "403":
          description: Forbidden - User does not have ADMIN role
          content:
            '*/*':
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                example: Internal server error occurred
...
  /api/v1/scheduled-trips/search:
    get:
      tags:
        - Scheduled Trip Management
      summary: Search Scheduled Trips
      description: "Search for scheduled trips based on origin, destination, and departure\
        \ date."
      operationId: SearchScheduledTrips
      parameters:
        - name: from
          in: query
          description: Origin location
          required: true
          schema:
            type: string
            example: Prague
        - name: to
          in: query
          description: Destination location
          required: true
          schema:
            type: string
            example: Vienna
        - name: date
          in: query
          description: Departure date in ISO format (yyyy-MM-dd)
          required: true
          schema:
            type: string
            format: date
            example: 2023-10-01
      requestBody:
        description: Tariff details for the trip
        content:
          application/json:
            schema:
              type: string
              enum:
                - ADULT
                - CHILD
                - STUDENT
                - SENIOR
        required: true
      responses:
        "200":
          description: List of scheduled trips matching the search criteria
          content:
            application/json:
              schema:
                type: array
                $ref: "#/components/schemas/ScheduledTripResponse"
                minProperties: 1
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                type: string
                example: Invalid date format or missing parameters
        "404":
          description: No scheduled trips found for the given criteria
          content:
            application/json:
              schema:
                type: string
                example: No scheduled trips found for the specified route and date
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                example: Internal server error occurred
...
components:
  schemas:
...
    ScheduledTripRequest:
      type: object
      description: Request object for scheduling a trip
      properties:
        busNumber:
          type: string
          description: Unique identifier for the bus
          example: BUS123
          minLength: 1
        origin:
          type: string
          description: Origin location of the trip
          example: Prague
          minLength: 1
        destination:
          type: string
          description: Destination location of the trip
          example: Vienna
          minLength: 1
        departureTime:
          type: string
          description: Departure time of the trip at departure's city time zone
          example: 08:30
        departureDay:
          type: array
          description: Days of the week when the trip is scheduled
          items:
            type: string
            enum:
              - MONDAY
              - TUESDAY
              - WEDNESDAY
              - THURSDAY
              - FRIDAY
              - SATURDAY
              - SUNDAY
          minItems: 1
          uniqueItems: true
        fromDate:
          type: string
          format: date
          description: Start date for the trip schedule in ISO format
          example: 2023-10-01
        toDate:
          type: string
          format: date
          description: End date for the trip schedule in ISO format
          example: 2023-12-31
      required:
        - busNumber
        - departureDay
        - departureTime
        - destination
        - fromDate
        - origin
        - toDate
...
    ScheduledTripResponse:
      type: object
      description: Response object for scheduled trip details
      properties:
        busNumber:
          type: string
          description: Unique identifier for the bus
          example: BUS123
        origin:
          type: string
          description: Origin of the trip
          example: Prague
        destination:
          type: string
          description: Destination of the trip
          example: Vienna
        departureDateTime:
          type: string
          format: date-time
          description: Departure date and time of the trip
          example: 2023-10-01T10:00:00Z
        arrivalDateTime:
          type: string
          format: date-time
          description: Arrival date and time of the trip
          example: 2023-10-01T12:00:00Z
        availableSeats:
          type: integer
          format: int32
          description: Number of available seats for the trip
          example: 30
          minimum: 0
        priceCzk:
          type: number
          description: Price of the trip in CZK
          example: 500.0
      required:
        - arrivalDateTime
        - availableSeats
        - busNumber
        - departureDateTime
        - destination
        - origin
        - priceCzk
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT